+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased] 2015-05-19
 ### Added
 - Functionality to allow Sales Invoices to be allocated to Projects and Tasks
+- Works order view from Projects
 
 ## [Unreleased][unreleased]
 ### Changed